requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
networkx==3.3
python-dotenv==1.0.1
matplotlib==3.8.4